extract macro logic to new class
